requirements.txt
confluent_kafka
psycopg2-binary
uvicorn
global-land-mask
numpy
fastapi